MTL-1309 Fix default value
The luksetcd module is supposed to key off of rd.luks, if it exists
then run the module. The default value though is 1, meaning if rd.luks
is removed from the cmdline that this will always run. Only by setting
rd.luks=0 would one disable the module, this is counter-intuitive. We
want to enable omitting the var to disable the creation. It isn't well
known that omitting it will default to 1.

This change forces the behavior to only run if rd.luks or rd.luks=1 is
set. Otherwise the omitting rd.luks or setting rd.luks=0 will have the
desired effect of not creating luks-etcd.