MTL-1560 Support Single Disk RAIDs (#13)
This fixes a bug that was discovered while using a single disk RAID (for
use in virtual or lab testing). This fixes an early exit, if multiple
candidates are found and the first one was insufficient the disk scanner
would exit early. This change ensures that we're always checking all
possible candidates.