VSHA-546 Iterate through all disks (#42)
Currently vshastav2 has an extra disk that shows up as first when all
disks are printed and sorted by size. This extra disk throws off the
`disk_offset`, which was used to skip/avoid disks consumed by the RAID.
Since this module does not iterate through disks, and relies on the disk
offset, it is prone to incorrectly choosing its disk if an
unused/non-rotofs disk is at the front.

The logic here has been replaced with a better scan method, each
available disk is iterated through and selected when it is large enough
and has no partitions. Since this module runs after the RAID module,
this will always skip the rootfs disks.